requirements.txt
numpy
python-dotenv
pandas
scipy
streamlit
plotly